refactor: 함수 네이밍 정리 및 파일 위치 재구성
