+++ src/app/App.tsx
@@ -2,15 +2,15 @@ import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 
 import AppRouter from '@/app/router';
-import { store } from '@/redux';
-import { AuthAndUserInfo, ScrollToTop } from '@/shared/hooks';
+import { store } from '@/app/store';
+import { useAuthAndUserInfo, useScrollToTop } from '@/shared/hooks';
 
 const App = () => {
+  useScrollToTop();
+  useAuthAndUserInfo();
   return (
     <BrowserRouter>
-      <ScrollToTop />
       <Provider store={store}>
-        <AuthAndUserInfo />
         <AppRouter />
       </Provider>
     </BrowserRouter>
