+++ src/pages/ui/HeroBanner/MacBookPro14/index.jsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
 
-import { productIdSlice } from '@/redux';
+import { productIdSlice } from '@/app/store';
 import { BREAKPOINTS } from '@/shared/assets/styled/breakPoints';
 import { FlexBetween } from '@/shared/assets/styled/CommonStyled';
