+++ src/shared/hooks/useAppDispatch.ts
@@ -1,5 +1,5 @@
 import { useDispatch } from 'react-redux';
 
-import type { AppDispatch } from '@/redux';
+import type { AppDispatch } from '@/app/store';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
