+++ src/shared/hooks/useAppSelector.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
 
-import type { RootState } from '@/redux';
+import type { RootState } from '@/app/store';
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
