+++ src/shared/layout/Header/index.jsx
@@ -1,11 +1,8 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { Link, useLocation } from 'react-router-dom';
 
-import { searchSlice } from '@/redux';
-import { LogoSVG } from '@/shared/assets/SVGicons';
-import { SearchSVG, CloseSVG } from '@/shared/assets/SVGicons';
-import { BagSVG } from '@/shared/assets/SVGicons';
-import { MenuSVG } from '@/shared/assets/SVGicons';
+import { searchSlice } from '@/app/store';
+import { BagSVG, CloseSVG, LogoSVG, MenuSVG, SearchSVG } from '@/shared/assets/SVGicons';
 import * as S from '@/shared/layout/Header/Header.styled';
 
 const Header = () => {
