+++ src/shared/lib/kebabToTitleCase.ts
@@ -1,0 +1,6 @@
+export const kebabToTitleCase = (word: string) => {
+  return word
+    .split('-')
+    .map((el) => el.charAt(0).toUpperCase() + el.slice(1))
+    .join(' ');
+};
